Implement async client simulation

+++ src/server/scripts/runner.sh
@@ -1,3 +1,12 @@
+#!/bin/bash
+
+set -e
+
+if [[ "$#" -ne 0 ]]; then
+    echo "Invalid number of arguments."
+    exit 1
+fi
+
 rm -fr ./build/ &&
 mkdir ./build/ &&
 cd ./build/ &&
